+++ client/src/components/editor/inputs/SourceInput.tsx
@@ -2,7 +2,7 @@ import React from "react";
 /** @jsx jsx */
 import { jsx, css } from "@emotion/core";
 import { EditorCompoenentInput } from "./@types";
-import SourceType from "../../../../types/source-type";
+import SourceType from "../../../types/source-type";
 
 const SourcePlaceholderMap: Record<SourceType, string> = {
   [SourceType.BOOK]: "책 정보...",
