+++ client/src/components/editor/inputs/TagInput.tsx
@@ -1,39 +1,41 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 /** @jsx jsx */
 import { jsx, css } from "@emotion/core";
-import TagList from "./TagList";
 import { Input } from "@material-ui/core";
+import TagList from "../../common/TagList";
+import { EditorCompoenentInput } from "./@types";
 
-const TagInputCss = css`
-  display: flex;
-  .tag-input {
-    margin-left: 1rem;
-  }
+const TagInputCss = css``;
 
-  & * {
-    font-size: 1rem;
-  }
-`;
-
-// deprecated
-const TagInput: React.FC<{
-  tags?: string[];
-  onChange: (tags: string[]) => void;
-}> = ({ tags = [], onChange }) => {
+const TagInput: EditorCompoenentInput<
+  "register" | "setValue" | "watch",
+  { defaultValue?: string[] }
+> = ({ register, setValue, watch, defaultValue = [] }) => {
   const [input, setInput] = useState("");
+  const tags: string[] = watch("tags", defaultValue);
   const handleTagOnKeypress: React.KeyboardEventHandler<HTMLInputElement> = (
     evt
   ) => {
     if (evt.key === "Enter" || evt.which === 13) {
       evt.preventDefault();
       const tag = input.trim();
-      if (tag.length > 0 && !tags.includes(tag)) onChange([...tags, tag]);
+      if (tag.length > 0 && !tags.includes(tag))
+        setValue("tags", [...tags, tag]);
       setInput("");
     }
   };
 
+  useEffect(() => {
+    register("tags");
+    setValue("tags", defaultValue);
+    // eslint-disable-next-line
+  }, []);
+
   const onTagClick = (tag: string) => {
-    onChange(tags.filter((t) => t !== tag));
+    setValue(
+      "tags",
+      tags.filter((t) => t !== tag)
+    );
   };
 
   return (
