+++ client/src/components/editor/inputs/__test__/SourceInput.test.tsx
@@ -4,7 +4,7 @@ import { fireEvent, render } from "@testing-library/react";
 import "mutationobserver-shim";
 
 import { useForm } from "react-hook-form";
-import SourceType from "../../../../../types/source-type";
+import SourceType from "../../../../types/source-type";
 import SourceInput from "../SourceInput";
 
 describe("<SourceInput />", () => {
