+++ client/src/components/renderer/renderer-component/MarkdownRenderer.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import MarkdownRenderer from "../Renderer/ArticleRenderer/ContentRenderer/MarkdownRenderer";
+import MarkdownRenderer from "./MarkdownRenderer";
 
 describe("MarkdownViewer", () => {
   const testMarkdown =
