+++ client/src/components/renderer/renderer-component/PlainRenderer.tsx
